requirements.txt
Keras==2.4.3
TensorFlow==2.4.1
Pandas==1.2.3
Numpy==1.19.5
Matplotlib==3.3.4
Pillow==8.1.0